+++ BeepBackend/Data/IArticleRepository.cs
@@ -14,9 +14,9 @@ namespace BeepBackend.Data
         Task<IEnumerable<ArticleGroup>> GetArticleGroups();
         Task<Article> LookupArticle(string barcode);
         Task<ArticleUserSetting> LookupArticleUserSettings(int articleId, int environmentId);
-        Task<int> GetArticleLifetime(string barcode, int environmentId);
+        Task<long> GetArticleLifetime(string barcode, int environmentId);
         Task<Article> SaveArticle(Article article, ArticleUserSetting userSettings);
-        Task<StockEntryValue> AddStockEntry(StockEntryValue entryValues, int usualLifetime);
+        Task<StockEntryValue> AddStockEntry(StockEntryValue entryValues, long usualLifetime);
         Task<PagedList<StockEntryValue>> GetStockEntries(int articleId, int environmentId, int page, int itemsPerPage);
         Task<DateTime> GetLastExpireDate(string barcode, int environmentId);
         Task<StockEntryValue> GetOldestStockEntryValue(string barcode, int environmentId);
